test/web-platform-tests/to-upstream/html/semantics/embedded-content/the-img-element/remove-attribute-src.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Disabled checkbox event handling</title>
<link rel="author" title="atsikov" href="mailto:alexey.tsikov@gmail.com">
<link rel=help href="https://html.spec.whatwg.org/#the-img-element">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>
<script>
"use strict";

async_test(t => {
  const grayPx = "data:image/gif;base64,R0lGODlhAQABAIAAAMLCwgAAACH5BAAAAAAALAAAAAABAAEAAAICRAEAOw==";
  const image = new window.Image();
  image.onload = () => {
    image.onload = () => {
      assert_true(false, "onload should not be called with removed src");
    };
    image.error = () => {
      assert_true(false, "error should not be called with removed src");
    };
    image.removeAttribute("src");
    assert_equals(image.width, 0);
    assert_equals(image.height, 0);
    assert_equals(image.src, "");
    assert_equals(image.currentSrc, "");
    t.done();
  };
  image.src = grayPx;
}, "img is cleared when src attribute is removed");
</script>
</body>
